move uuid package to internal, remove unused functions
