+++ api/api_test.go
@@ -20,7 +20,7 @@ import (
 	"github.com/Ajnasz/sekret.link/internal/durable"
 	"github.com/Ajnasz/sekret.link/internal/models"
 	"github.com/Ajnasz/sekret.link/internal/services"
-	"github.com/Ajnasz/sekret.link/uuid"
+	"github.com/Ajnasz/sekret.link/internal/uuid"
 )
 
 func NewHandlerConfig(db *sql.DB) HandlerConfig {
