--- uuid/uuid_test.go
@@ -1,29 +1,0 @@
-package uuid
-
-import "testing"
-
-func TestGetUUIDFromPath(t *testing.T) {
-	testCases := []struct {
-		Name     string
-		Value    string
-		Expected string
-	}{
-		{
-			"simple uuid",
-			"/3f356f6c-c8b1-4b48-8243-aa04d07b8873",
-			"3f356f6c-c8b1-4b48-8243-aa04d07b8873",
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.Name, func(t *testing.T) {
-			actual, err := GetUUIDFromPath(testCase.Value)
-			if err != nil {
-				t.Fatal(err)
-			}
-			if testCase.Expected != actual {
-				t.Errorf("expected: %q, actual: %q", testCase.Expected, actual)
-			}
-		})
-	}
-}
